ai_agent_service/requirements.txt
fastapi==0.95.1
uvicorn==0.20.0
redis==4.5.1
pika==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.1
python-dotenv==1.0.0
httpx==0.24.1
locust==2.19.0
python-multipart==0.0.6
langchain==0.0.256
openai==0.27.8
pymongo==4.3.3
motor==3.1.1
